+++ Changelog.md
@@ -1,5 +1,10 @@
 # PHBackup changelog
 
+## 1.6.2
+* Fixed a typo in default pre_script leading to mariabackup error
+* Updated pre-scripts for all hosts in group id 1 (servers)
+* Added /etc/phbackup/functions.custom.php
+
 ## 1.6.1
 
 * Fixed Rsync status handling